+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/DependencyManagementPluginKitTest.groovy
@@ -145,7 +145,7 @@ class DependencyManagementPluginKitTest extends AbstractKitTest {
                 compileOnly 'junit:junit'
             }  
 
-            pom {
+            maven.pom {
                 developers {
                     developer {
                         id 'jdoe'