+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/MultipleDevelopersKitTest.groovy
@@ -28,7 +28,7 @@ class MultipleDevelopersKitTest extends AbstractKitTest {
                 }
             }
 
-            pom {
+            maven.pom {
                 developers {
                     developer {
                         id "dev1"