+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/OptionalDepsKitTest.groovy
@@ -42,7 +42,7 @@ class OptionalDepsKitTest extends AbstractKitTest {
                 }
             }
 
-            pom {
+            maven.pom {
                 developers {
                     developer {
                         id "dev"