+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/WarCompatibilityTest.groovy
@@ -37,7 +37,7 @@ class WarCompatibilityTest extends AbstractKitTest {
                 }
             }
 
-            pom {
+            maven.pom {
                 developers {
                     developer {
                         id "dev"